// apps/console/src/__generated__/core/AccessReviewPageFragment.graphql.ts
export type AccessReviewPageFragment$data = {
  readonly campaigns: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"AccessReviewCampaignRowFragment">;
      };
    }>;
  };
  readonly id: string;
  readonly " $fragmentType": "AccessReviewPageFragment";
};
export type AccessReviewPageFragment$key = {
  readonly " $data"?: AccessReviewPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"AccessReviewPageFragment">;
};